requirements.txt
numpy==1.26.4
pillow==10.2.0
tensorflow-gpu==2.10.0
scikit-learn==1.5.0
seaborn==0.13.2
matplotlib==3.9.0